% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ACTIONetExperiment-class}
\alias{ACTIONetExperiment-class}
\alias{ACTIONetExperiment}
\title{An extension of the SummarizedExperiment class to store
the results of ACTIONet method}
\usage{
ACTIONetExperiment(
  ...,
  rowNets = S4Vectors::SimpleList(),
  colNets = S4Vectors::SimpleList(),
  rowMaps = S4Vectors::SimpleList(),
  colMaps = S4Vectors::SimpleList()
)
}
\arguments{
\item{...}{SummarizedExperiment and SummarizedExperiment components}

\item{rowNets, colNets}{gene-gene and cell-cell networks, respectively}

\item{rowMaps, colMaps}{Factorization results (W and H matrices)}
}
\value{
an ACTIONetExperiment (ACE) object

An ACTIONetExperiment (ACE) object, derived from SummarizedExperiment, with additional slots to store ACTIONet results
}
\description{
An extension of the SummarizedExperiment class to store
the results of ACTIONet method

Creates an ACTIONetExperiment (ACE) object
}
\section{Slots}{

\describe{
\item{\code{rowNets,colNets}}{gene-gene and cell-cell networks, respectively}

\item{\code{rowMaps,colMaps}}{Factorization results (W and H matrices)}
}}

