% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate.archetypes.using.markers}
\alias{annotate.archetypes.using.markers}
\title{Annotate clusters using known marker genes
(It uses permutation test on cluster specificity scores)}
\usage{
annotate.archetypes.using.markers(
  ace,
  markers,
  rand_sample_no = 1000,
  significance_slot = "unified_feature_specificity"
)
}
\arguments{
\item{ace}{ACTIONet output object}

\item{markers}{A list of lists (each a set of markers for a given cell type)}

\item{rand_sample_no}{Number of random permutations (default=1000)}
}
\value{
A named list: \itemize{
\item Label: Inferred archetype labels
\item Confidence: Confidence of inferred labels
\item Enrichment: Full enrichment matrix
}
}
\description{
Annotate clusters using known marker genes
(It uses permutation test on cluster specificity scores)
}
\examples{
data('curatedMarkers_human') # pre-packaged in ACTIONet
markers = curatedMarkers_human$Blood$PBMC$Monaco2019.12celltypes$marker.genes
arch.annot = annotate.archetypes.using.markers(ace, markers = markers)
}
