% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate.cells.from.archetypes.using.markers}
\alias{annotate.cells.from.archetypes.using.markers}
\title{Annotates cells by interpolating marker-based archetype annotations}
\usage{
annotate.cells.from.archetypes.using.markers(
  ace,
  markers,
  rand_sample_no = 1000,
  unified_suffix = "unified"
)
}
\arguments{
\item{ace}{ACTIONet output object}

\item{markers}{A list of lists (each a set of markers for a given cell type)}

\item{rand_sample_no}{Number of random permutations (default=1000)}
}
\value{
A named list: \itemize{
\item Label: Inferred archetype labels
\item Confidence: Confidence of inferred labels
\item Enrichment: Full enrichment matrix
}
}
\description{
Annotates cells by interpolating marker-based archetype annotations
}
\examples{

data('curatedMarkers_human') # pre-packaged in ACTIONet
marker_set = curatedMarkers_human$Blood$PBMC$Monaco2019.12celltypes$marker.genes
cell.annotations = annotate.cells.from.archetypes.using.markers(ace, markers)
labels = cell.annotations$Labels
}
