% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate.cells.using.markers}
\alias{annotate.cells.using.markers}
\title{Infer cell annotations from imputed gene expression for all cells.}
\usage{
annotate.cells.using.markers(
  ace,
  markers,
  features_use = NULL,
  alpha_val = 0.9,
  thread_no = 0,
  net_slot = "ACTIONet",
  assay_name = "logcounts",
  max_iter = 5
)
}
\arguments{
\item{ace}{ACTIONetExperiment object}

\item{markers}{A named list of marker genes.}

\item{features_use}{A vector of features of length NROW(ace) or the name of a column of rowData(ace) containing the genes given in 'markers'.}

\item{alpha_val}{Random-walk parameter for gene imputation.}

\item{thread_no}{Number of parallel threads used for gene imputation.}

\item{net_slot}{Name of slot in colNets(ace) containing the network to use for gene expression imputation (default="ACTIONet").}

\item{assay_name}{Name of assay for which to impute gene expression (default="logcounts").}
}
\value{
A named list: \itemize{
\item Label: Inferred cell type labels
\item Confidence: Confidence of inferred labels
\item Enrichment: Cell type score matrix.
}
}
\description{
Infer cell annotations from imputed gene expression for all cells.
}
\examples{
data('curatedMarkers_human') # pre-packaged in ACTIONet
markers = curatedMarkers_human$Blood$PBMC$Monaco2019.12celltypes$marker.genes
annots = annotate.cells.using.markers(ace, markers = markers)
plot.ACTIONet(ace, annots$Label, annots$Confidence)
}
