% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{annotate.clusters.using.labels}
\alias{annotate.clusters.using.labels}
\title{Annotate clusters using prior cell annotations
(It uses Fisher's exact test for computing overlaps -- approximate HGT is used)}
\usage{
annotate.clusters.using.labels(ace, clusters, labels)
}
\arguments{
\item{ace}{ACTIONet output object}

\item{clusters}{Cluster}

\item{labels}{Annotation of interest (clusters, celltypes, etc.) to test enrichment}
}
\value{
A named list: \itemize{
\item Labels: Inferred archetype labels
\item Labels.confidence: Confidence of inferred labels
\item Enrichment: Full enrichment matrix
}
}
\description{
Annotate clusters using prior cell annotations
(It uses Fisher's exact test for computing overlaps -- approximate HGT is used)
}
\examples{
arch.annot = annotate.clusters.using.labels(ace, ace$clusters, sce$celltypes)
}
