% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{annotate.profile.using.markers}
\alias{annotate.profile.using.markers}
\title{Annotate arbitary feature score matrix using known marker genes
(It uses permutation test on cluster specificity scores)}
\usage{
annotate.profile.using.markers(
  feature.scores,
  marker.genes,
  rand.sample.no = 1000
)
}
\arguments{
\item{feature.scores}{An arbitrary matrix with rows corresponding to features and columns to any given annotation/grouping of cells}

\item{marker.genes}{A list of lists (each a set of markers for a given cell type)}

\item{rand.sample.no}{Number of random permutations (default=1000)}
}
\value{
A named list: \itemize{
\item Labels: Inferred archetype labels
\item Labels.confidence: Confidence of inferred labels
\item Enrichment: Full enrichment matrix
}
}
\description{
Annotate arbitary feature score matrix using known marker genes
(It uses permutation test on cluster specificity scores)
}
\examples{
data('curatedMarkers_human') # pre-packaged in ACTIONet
marker.genes = curatedMarkers_human$Blood$PBMC$Monaco2019.12celltypes$marker.genes
arch.annot = annotate.profile.using.markers(my.gene.scores.profile, marker.genes = marker.genes)
}
