% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_ace.R
\name{as.ACTIONetExperiment}
\alias{as.ACTIONetExperiment}
\title{Converts a SummarizedExperiment, RangedSummarizedExperiment, or SingleCellExperiment object to an ACTIONetExperiment object.}
\usage{
as.ACTIONetExperiment(object)
}
\arguments{
\item{object}{SummarizedExperiment, RangedSummarizedExperiment, or SingleCellExperiment object}
}
\description{
Converts a SummarizedExperiment, RangedSummarizedExperiment, or SingleCellExperiment object to an ACTIONetExperiment object.
}
