% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_ace.R
\name{as.SingleCellExperiment}
\alias{as.SingleCellExperiment}
\title{Converts an ACTIONetExperiment object to a SingleCellExperiment object.}
\usage{
as.SingleCellExperiment(ace)
}
\arguments{
\item{ace}{ACTIONetExperiment object}
}
\description{
Converts an ACTIONetExperiment object to a SingleCellExperiment object.
}
