% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_ace.R
\name{as.SummarizedExperiment}
\alias{as.SummarizedExperiment}
\title{Converts an ACTIONetExperiment method to a SummarizedExperiment object.}
\usage{
as.SummarizedExperiment(ace)
}
\arguments{
\item{ace}{ACTIONetExperiment object}
}
\description{
Converts an ACTIONetExperiment method to a SummarizedExperiment object.
}
