% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{assess.geneset.enrichment.gProfiler}
\alias{assess.geneset.enrichment.gProfiler}
\title{Performs geneset enrichment analysis on a given set of genes
(It uses permutation test on cluster specificity scores)}
\usage{
assess.geneset.enrichment.gProfiler(
  genes,
  category = c("GO:BP", "REAC", "KEGG"),
  organism = "hsapiens",
  top.terms = 10,
  col = "tomato"
)
}
\arguments{
\item{genes}{Set of genes}

\item{category}{List of functional categories to analyze (default: c('GO:BP', 'REAC', 'KEGG'))}

\item{organism}{Species name (default = 'hsapiens')}

\item{top.terms}{Number of terms to sho}

\item{col}{Color of the barplot}
}
\value{
Bar plot of the top-ranked terms
}
\description{
Performs geneset enrichment analysis on a given set of genes
(It uses permutation test on cluster specificity scores)
}
\examples{
geneset.enrichment.gProfiler(my_genes)
}
