% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{assess.TF.activities.from.scores}
\alias{assess.TF.activities.from.scores}
\title{Compute the activity score of each TF based on the observed activity of its targets.
(Similar to ChEA3; it is based on a meta-analysis using 4 separate datasets)}
\usage{
assess.TF.activities.from.scores(scores)
}
\arguments{
\item{scores}{Matrix of genes x cell type/state}
}
\value{
Matrix of TF x cell type/state indicating inferred TF activity scores
}
\description{
Compute the activity score of each TF based on the observed activity of its targets.
(Similar to ChEA3; it is based on a meta-analysis using 4 separate datasets)
}
\examples{
scores = rowMaps(ace)$unified_feature_specificity
TF.scores = assess.TF.activities.from.scores(scores)
}
