% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CellMarkerDB_human_markers}
\alias{CellMarkerDB_human_markers}
\title{Pre-processed cell type markers from the CellMarker database (Human)}
\format{
:
  A list of marker genes for different cell types
}
\source{
\url{http://biocc.hrbmu.edu.cn/CellMarker/}
}
\usage{
CellMarkerDB_human_markers
}
\description{
Pre-processed cell type markers from the CellMarker database (Human)
}
\keyword{datasets}
