% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChEA3plusDB}
\alias{ChEA3plusDB}
\title{Collection of TF-TG assocations}
\format{
:
  A list of TF-TG associations
}
\source{
\url{https://amp.pharm.mssm.edu/chea3/}
}
\usage{
ChEA3plusDB
}
\description{
Collection of TF-TG assocations
}
\keyword{datasets}
