% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{colEmbeddings<-}
\alias{colEmbeddings<-}
\title{Set column embeddings}
\usage{
colEmbeddings(object, ...) <- value
}
\arguments{
\item{object}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Set column embeddings
}
