% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{colMaps<-}
\alias{colMaps<-}
\title{Set colMaps}
\usage{
colMaps(object, ...) <- value
}
\arguments{
\item{object}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Set colMaps
}
