% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/get_ace.R, R/set_ace.R
\name{colMapTypes}
\alias{colMapTypes}
\alias{colMapTypes,ACTIONetExperiment-method}
\alias{colMapMeta,ACTIONetExperiment-method}
\alias{colMapTypes<-,ACTIONetExperiment-method}
\title{Get colMapTypes}
\usage{
colMapTypes(object, ...)

\S4method{colMapTypes}{ACTIONetExperiment}(object, all = TRUE)

\S4method{colMapMeta}{ACTIONetExperiment}(object, all = TRUE)

\S4method{colMapTypes}{ACTIONetExperiment}(object) <- value
}
\arguments{
\item{object}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\value{
List of types

List of types

List of matrices
}
\description{
Get colMapTypes

Get column-associated factor types

Get column-associated factor metadata

Set column-associated factor annotations
}
