% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{colNets<-}
\alias{colNets<-}
\title{Set colNets}
\usage{
colNets(object, ...) <- value
}
\arguments{
\item{object}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Set colNets
}
