% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublet.R
\name{compute.ACTIONet.doublet.score}
\alias{compute.ACTIONet.doublet.score}
\title{Computes quality scores for cells}
\usage{
compute.ACTIONet.doublet.score(
  ace,
  z.threshold = 1,
  combination_method = "mean"
)
}
\arguments{
\item{ace}{ACTIONet output}
}
\value{
Updated ace object with dblscore.expression, dblscore.centrality, and dblscore.combined scores
}
\description{
Computes quality scores for cells
}
\examples{
ace = compute.ACTIONet.doublet.score(ace)
}
