% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{compute.cluster.feature.specificity}
\alias{compute.cluster.feature.specificity}
\title{Computes feature (i.e. gene) specificity scores for each cluster}
\usage{
compute.cluster.feature.specificity(
  ace,
  clusters,
  output_slot,
  assay_name = "logcounts"
)
}
\arguments{
\item{ace}{ACTIONet output object}

\item{clusters}{Cluster}

\item{output_slot}{Name of the output in rowMaps(ace) to store results}

\item{renormalize.logcounts.slot}{Name of the new assay with updated logcounts adjusted using archetypes
Typically it is either 'logcounts' or 'logcounts'}
}
\value{
`ACE` object with specificity scores of each cluster added to rowMaps(ace) as a matrix with name defined by output_slot
}
\description{
Computes feature (i.e. gene) specificity scores for each cluster
}
\examples{
ace = compute.cluster.feature.specificity(ace, ace$clusters, 'cluster_specificity_scores')
}
