% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_core_number}
\alias{compute_core_number}
\title{Compute coreness of graph vertices}
\usage{
compute_core_number(G)
}
\arguments{
\item{G}{Input graph}
}
\value{
cn core-number of each graph node
}
\description{
Compute coreness of graph vertices
}
\examples{
G = colNets(ace)$ACTIONet
cn = compute_core_number(G)
}
