% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decode_ids}
\alias{decode_ids}
\title{Decrypts a set of given encrypted ids}
\usage{
decode_ids(encoded_ids, pass)
}
\arguments{
\item{encoded_ids}{List of encrypted string ids}

\item{pass}{Pass phrase to use for decryption}
}
\value{
A string array of decrypted ids
}
\description{
Decrypts a set of given encrypted ids
}
\examples{
ids = decode_ids(encoded.ids)
}
