% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DiseaseDB}
\alias{DiseaseDB}
\title{Collection of Disease-associated genesets}
\format{
:
  A list of genesets
}
\source{
\url{https://amp.pharm.mssm.edu/Enrichr/}
}
\usage{
DiseaseDB
}
\description{
Collection of Disease-associated genesets
}
\keyword{datasets}
