% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{encode_ids}
\alias{encode_ids}
\title{Encrypts a set of given input ids}
\usage{
encode_ids(ids, pass)
}
\arguments{
\item{ids}{List of input string ids}

\item{pass}{Pass phrase to use for encryption}
}
\value{
A string array of encoded ids
}
\description{
Encrypts a set of given input ids
}
\examples{
encoded.ids = encode_ids(colnames(sce))
}
