% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ace.R
\name{filter.ace.by.attr}
\alias{filter.ace.by.attr}
\title{Filter columns and rows of `ACTIONetExperiment` or `SummarizedExperiment` object by column attribute.}
\usage{
\method{filter}{ace.by.attr}(
  ace,
  by,
  assay_name = "counts",
  min_cells_per_feat = NULL,
  min_feats_per_cell = NULL,
  min_umis_per_cell = NULL,
  max_umis_per_cell = NULL
)
}
\description{
Filter columns and rows of `ACTIONetExperiment` or `SummarizedExperiment` object by column attribute.
}
