% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ace.R
\name{filter.ace}
\alias{filter.ace}
\title{Filter columns and rows of `ACTIONetExperiment` or `SummarizedExperiment`-like object.}
\usage{
\method{filter}{ace}(
  ace,
  assay_name = "counts",
  min_cells_per_feat = NULL,
  min_feats_per_cell = NULL,
  min_umis_per_cell = NULL,
  max_umis_per_cell = NULL,
  return_fil_ace = TRUE
)
}
\description{
Filter columns and rows of `ACTIONetExperiment` or `SummarizedExperiment`-like object.
}
