% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gProfilerDB_human}
\alias{gProfilerDB_human}
\title{Pre-processed genesets from the gProfiler database (Human)}
\format{
:
  SYMBOL: A list of gene x geneset indicator matrices in the sparseMatrix format (rownames are gene symbols)
  ENSEMBL: A list of gene x geneset indicator matrices in the sparseMatrix format (rownames are ENSEMBL gene IDs)
}
\source{
\url{https://biit.cs.ut.ee/gprofiler}
}
\usage{
gProfilerDB_human
}
\description{
Pre-processed genesets from the gProfiler database (Human)
}
\keyword{datasets}
