% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gProfilerDB_mouse}
\alias{gProfilerDB_mouse}
\title{Pre-processed genesets from the gProfiler database (Mouse)}
\format{
:
  SYMBOL: A list of gene x geneset indicator matrices in the sparseMatrix format (rownames are gene symbols)
  ENSEMBL: A list of gene x geneset indicator matrices in the sparseMatrix format (rownames are ENSEMBL gene IDs)
}
\source{
\url{https://biit.cs.ut.ee/gprofiler}
}
\usage{
gProfilerDB_mouse
}
\description{
Pre-processed genesets from the gProfiler database (Mouse)
}
\keyword{datasets}
