% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{HDBSCAN.clustering}
\alias{HDBSCAN.clustering}
\title{A wrapper function For HDBSCAN algorithm applied to an ACE object}
\usage{
HDBSCAN.clustering(
  ace,
  minPoints = 30,
  minClusterSize = 30,
  archetype.slot = "H_unified"
)
}
\arguments{
\item{ace}{Input results to be clustered
minPoints, minClusterSize HDBSCAN parameters (default = 30,30)
archetype.slot Slot of archeypte to use for clustering (default='H_unified');}
}
\value{
clusters
}
\description{
A wrapper function For HDBSCAN algorithm applied to an ACE object
}
\examples{
clusters = HDBSCAN.clustering(ace)
plot.ACTIONet(ace, clusters)
}
