% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hs2Mm}
\alias{Hs2Mm}
\title{Human-Mouse orthologs}
\format{
:
  data.frame with two columns corresponding to human and mouse genes, respectively
}
\source{
\url{http://www.informatics.jax.org/downloads/reports/HMD_HumanPhenotype.rpt}
}
\usage{
Hs2Mm
}
\description{
Human-Mouse orthologs
}
\keyword{datasets}
