% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import.ace.from.10X.h5}
\alias{import.ace.from.10X.h5}
\title{A simple wrapper to import data from a 10X experiment folder (v2 or v3) and constructs an `SingleCellExeriment` object}
\usage{
import.ace.from.10X.h5(
  fname,
  version = 3,
  genome = NULL,
  use.names = TRUE,
  prefilter = FALSE,
  ...
)
}
\arguments{
\item{fname}{Input HDF5 file}

\item{version}{2 or 3}

\item{genome}{genome to import. Default is to import the first (if many)}

\item{prefilter}{Whether to prefilter genes/cells based on the counts.mat}

\item{min.cell.frac.per.gene}{Minimum fraction of cells capturing a gene for it to be retained, if prefilter=TRUE (default=0.005)}

\item{min.genes.per.cell}{Minimum number of required captured genes per cell, if prefilter=TRUE (default=500)}
}
\value{
`SingleCellExeriment` object
}
\description{
A simple wrapper to import data from a 10X experiment folder (v2 or v3) and constructs an `SingleCellExeriment` object
}
\examples{
ace = import.ace.from.10X.h5(fname = fname, prefilter=TRUE)
}
