% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import.ace.from.Seurat}
\alias{import.ace.from.Seurat}
\title{Constructs an `SingleCellExeriment` object from a Seurat object
Please refer to: https://satijalab.org/seurat/v3.0/conversion_vignette.html}
\usage{
import.ace.from.Seurat(Seurat.obj)
}
\arguments{
\item{Seurat.obj}{Seurat object}
}
\value{
`SingleCellExeriment` object
}
\description{
Constructs an `SingleCellExeriment` object from a Seurat object
Please refer to: https://satijalab.org/seurat/v3.0/conversion_vignette.html
}
\examples{
ace = import.ace.from.Seurat(file_name)
}
