% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import.ace.from.table}
\alias{import.ace.from.table}
\title{Constructs an `SingleCellExeriment` object from a full count matrix file}
\usage{
import.ace.from.table(fname, sep = "\\t", prefilter = FALSE, ...)
}
\arguments{
\item{fname}{Full path to the count matrix file}

\item{sep}{Column-separator used in count matrix file (default='\\t')}

\item{prefilter}{Whether to prefilter genes/cells based on the counts.mat}

\item{min.cell.frac.per.gene}{Minimum fraction of cells capturing a gene for it to be retained, if prefilter=TRUE (default=0.005)}

\item{min.genes.per.cell}{Minimum number of required captured genes per cell, if prefilter=TRUE (default=500)}
}
\value{
`SingleCellExeriment` object
}
\description{
Constructs an `SingleCellExeriment` object from a full count matrix file
}
\examples{
ace = import.ace.from.table(file_name, prefilter=TRUE)
}
