% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute.genes.using.ACTIONet}
\alias{impute.genes.using.ACTIONet}
\title{Gene expression imputation using network diffusion.}
\usage{
impute.genes.using.ACTIONet(
  ace,
  genes,
  features_use = NULL,
  alpha_val = 0.85,
  thread_no = 0,
  diffusion_iters = 5,
  assay_name = "logcounts"
)
}
\arguments{
\item{ace}{Input results to be clustered
(alternatively it can be the ACTIONet igraph object)}

\item{genes}{The list of genes to perform imputation for.}

\item{alpha_val}{Depth of diffusion between (0, 1).
The larger it is, the deeper the diffusion, which results in less nonzeros (default = 0.85).}

\item{thread_no}{Number of parallel threads}

\item{diffusion_iters}{Number of diffusion iterations (default = 5)}

\item{assay_name}{Slot in the ace object with normalized counts.}
}
\value{
Imputed gene expression matrix. Column names are set with imputed genes names and rows are cells.
}
\description{
Gene expression imputation using network diffusion.
}
\examples{
imputed.genes = impute.genes.using.ACTIONet(ace, c('CD14', 'CD19', 'CD3G'))
plot.ACTIONet.gradient(ace, imputed.genes[, 1])
}
