% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{IRLB_SVD_full}
\alias{IRLB_SVD_full}
\title{Computes SVD decomposition}
\usage{
IRLB_SVD_full(A, dim, iters = 1000L, seed = 0L, verbose = 1L)
}
\arguments{
\item{A}{Input matrix ("sparseMatrix")}

\item{dim}{Dimension of SVD decomposition}

\item{iters}{Number of iterations (default=5)}

\item{seed}{Random seed (default=0)}
}
\value{
A named list with U, sigma, and V components
}
\description{
This is direct implementation of the randomized SVD algorithm:
From: IRLBA R Package
}
\examples{
A = randn(100, 20)
SVD.out = IRLBA_SVD_full(A, dim = 2)
U = SVD.out$U
}
