% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{Leiden.clustering}
\alias{Leiden.clustering}
\title{A wrapper function For Leiden algorithm applied to an ACE object}
\usage{
Leiden.clustering(
  ace,
  resolution_parameter = 1,
  net.slot = "ACTIONet",
  init.slot = "assigned_archetype",
  seed = 0,
  postprocess = T,
  PP_lambda = 0,
  PP_iters = 3,
  PP_sig_threshold = 3
)
}
\arguments{
\item{ace}{Input results to be clustered}

\item{resolution_parameter}{Resolution of the clustering.
The higher the resolution, the more clusters we will get (default=0.5).}

\item{seed}{Random seed}

\item{arch.init}{Whether to use archetype-assignments to initialize clustering (default=TRUE)}
}
\value{
clusters
}
\description{
A wrapper function For Leiden algorithm applied to an ACE object
}
\examples{
clusters = Leiden.clustering(ace)
plot.ACTIONet(ace, clusters)
}
