% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{map.cell.scores.from.archetype.enrichment}
\alias{map.cell.scores.from.archetype.enrichment}
\title{Interpolates cell scores from archetype enrichment matrix}
\usage{
map.cell.scores.from.archetype.enrichment(
  ace,
  enrichment_mat,
  normalize = FALSE,
  H.slot = "H_unified"
)
}
\arguments{
\item{ace}{ACTIONet output object}

\item{enrichment_mat}{Enrichment matrix with rows corresponding to archetypes and columns to an arbitrary annotation}

\item{normalize}{If TRUE, enrichment matrix will be first doubly-normalized}
}
\value{
Enrichment map of size cell x annotation
}
\description{
Interpolates cell scores from archetype enrichment matrix
}
\examples{

data('curatedMarkers_human') # pre-packaged in ACTIONet
marker_set = curatedMarkers_human$Blood$PBMC$Monaco2019.12celltypes$marker.genes
arch.annot = annotate.archetypes.using.markers(ace, markers = markers)
enrichment.mat = arch.annot$enrichment
cell.enrichment.mat = map.cell.scores.from.archetype.enrichment(ace, enrichment.mat)
cell.assignments = colnames(cell.enrichment.mat)[apply(cell.enrichment.mat, 1, which.max)]
}
