% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nanoStringDB_mouse}
\alias{nanoStringDB_mouse}
\title{Pre-processed genesets from the nanoString database (Mouse)}
\format{
:
  A list of gene x geneset indicator matrices in the sparseMatrix format (rownames are gene symbols)
}
\source{
\url{https://www.nanostring.com/}
}
\usage{
nanoStringDB_mouse
}
\description{
Pre-processed genesets from the nanoString database (Mouse)
}
\keyword{datasets}
