% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NetDBSCAN}
\alias{NetDBSCAN}
\title{Computes disjoint clusters for vertices of G.
(It uses an adjusted DBSCAN procedure)}
\usage{
NetDBSCAN(G, minPts = 10L, eps = 0.5, alpha = 0.85)
}
\arguments{
\item{G}{Adjacency matrix of the input graph}

\item{minPts, }{eps DBSCAN parameters}

\item{alpha}{Diffusion parameter for initial node ordering}
}
\value{
Matrix of log-pvalues
}
\description{
Computes disjoint clusters for vertices of G.
(It uses an adjusted DBSCAN procedure)
}
\examples{
G = colNets(ace)$ACTIONet
clusters = NetDBSCAN(G)
}
