% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PanglaoDB_human_markers}
\alias{PanglaoDB_human_markers}
\title{Pre-processed cell type markers from the PanglaoDB database (Human)}
\format{
:
  A list of marker genes for different cell types
}
\source{
\url{https://panglaodb.se/}
}
\usage{
PanglaoDB_human_markers
}
\description{
Pre-processed cell type markers from the PanglaoDB database (Human)
}
\keyword{datasets}
