% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PanglaoDB_mouse_markers}
\alias{PanglaoDB_mouse_markers}
\title{Pre-processed cell type markers from the PanglaoDB database (Mouse)}
\format{
:
  A list of marker genes for different cell types
}
\source{
\url{https://panglaodb.se/}
}
\usage{
PanglaoDB_mouse_markers
}
\description{
Pre-processed cell type markers from the PanglaoDB database (Mouse)
}
\keyword{datasets}
