% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.ACTIONet.3D}
\alias{plot.ACTIONet.3D}
\title{Main ACTIONet 3D plotting functions}
\usage{
\method{plot}{ACTIONet.3D}(
  ace,
  labels = NULL,
  transparency.attr = NULL,
  trans.z.threshold = -1,
  trans.fact = 1,
  node_size = 1,
  CPal = CPal20,
  coordinate_slot = "ACTIONet3D"
)
}
\arguments{
\item{ace}{ACTIONet output object}

\item{labels}{Annotation of interest (clusters, celltypes, etc.) to be projected on the ACTIONet plot}

\item{transparency.attr}{Additional continuous attribute to project onto the transparency of nodes}

\item{trans.z.threshold, }{trans.fact Control the effect of transparency mapping}

\item{node_size}{Size of nodes in the ACTIONet plot}

\item{CPal}{Color palette (named vector or a name for a given known palette)}

\item{coordinate_slot}{Entry in colMaps(ace) containing the plot coordinates (default:'ACTIONet2D')}

\item{title}{Main title of the plot}
}
\value{
Visualized ACTIONet
}
\description{
Main ACTIONet 3D plotting functions
}
\examples{
ace = run.ACTIONet(sce)
plot.ACTIONet.3D(ace, ace$assigned_archetype, transparency.attr = ace$node_centrality)
}
