% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.ACTIONet.gene.view}
\alias{plot.ACTIONet.gene.view}
\title{Plots projection of genes onto the ACTIONet plot}
\usage{
\method{plot}{ACTIONet.gene.view}(
  ace,
  top_genes = 5,
  CPal = NULL,
  blacklist_pattern = "\\\\.|^RPL|^RPS|^MRP|^MT-|^MT|MALAT1|B2M|GAPDH",
  title = "",
  label_size = 0.8,
  renormalize = FALSE
)
}
\arguments{
\item{ace}{ACTIONet output}

\item{top_genes}{Number of genes to return}

\item{CPal}{Color palette to use}

\item{blacklist_pattern}{List of genes to filter-out}

\item{title}{Main title of the plot}

\item{reorder_columns}{Whether to optimally re-order columns of the enrichment table}
}
\value{
Featur view
}
\description{
Plots projection of genes onto the ACTIONet plot
}
\examples{
plot.ACTIONet.gene.view(ace, 5)
}
