% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.ACTIONet.gradient}
\alias{plot.ACTIONet.gradient}
\title{Projects a given continuous score on the ACTIONet plot}
\usage{
\method{plot}{ACTIONet.gradient}(
  ace,
  x,
  transparency.attr = NULL,
  trans.z.threshold = -0.5,
  trans.fact = 3,
  node_size = 0.1,
  CPal = "magma",
  title = "",
  alpha_val = 0.85,
  nonparameteric = FALSE,
  coordinate_slot = "ACTIONet2D"
)
}
\arguments{
\item{ace}{ACTIONet output object}

\item{x}{Score vector}

\item{transparency.attr}{Additional continuous attribute to project onto the transparency of nodes}

\item{trans.z.threshold, }{trans.fact Control the effect of transparency mapping}

\item{node_size}{Size of nodes in the ACTIONet plot}

\item{CPal}{Color palette (named vector or a name for a given known palette)}

\item{alpha_val}{Between [0, 1]. If it is greater than 0, smoothing of scores would be performed}

\item{coordinate_slot}{Entry in colMaps(ace) containing the plot coordinates (default:'ACTIONet2D')}
}
\value{
Visualized ACTIONet with projected scores
}
\description{
Projects a given continuous score on the ACTIONet plot
}
\examples{
ace = run.ACTIONet(ace)
x = logcounts(ace)['CD14', ]
plot.ACTIONet.gradient(ace, x, transparency.attr = ace$node_centrality)
}
