% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{plot.ACTIONet.gradient.test}
\alias{plot.ACTIONet.gradient.test}
\title{Plots gradient of (potentially imputed) values on ACTIONet scatter plot.}
\usage{
\method{plot}{ACTIONet.gradient.test}(
  ace,
  x,
  alpha_val = 0.85,
  log_scale = FALSE,
  nonparameteric = FALSE,
  trans_attr = NULL,
  trans_fac = 1.5,
  trans_th = -0.5,
  point_size = 1,
  stroke_size = point_size * 0.1,
  stroke_contrast_fac = 0.1,
  grad_palette = "magma",
  net_attr = "ACTIONet",
  coordinate_attr = "ACTIONet2D"
)
}
\arguments{
\item{ace}{'ACTIONetExperiment' object}

\item{x}{Numeric vector of length NCOL(ace).}

\item{alpha_val}{Smoothing parameter for PageRank imputation of 'x'. No imputation if 'alpha_val=0' (default:0.85).}

\item{log_scale}{Logical value for whether to log-scale values of 'x' (default:'FALSE').}

\item{nonparameteric}{If 'FALSE', values of 'x' are used as breaks for color gradient. If 'TRUE' the ranks of the values of 'x' are used instead  (default:'FALSE').}

\item{trans_attr}{Numeric vector of length NROW(ace) or colname of 'colData(ace)' used to compute point transparency. Smaller values are more transparent.}

\item{trans_fac}{Transparency modifier (default:1.5).}

\item{trans_th}{Minimum Z-score for which points with 'scale(trans_attr) < trans_th' are masked (default:-0.5).}

\item{point_size}{Size of points in ggplot (default:1).}

\item{stroke_size}{Size of points outline (stroke) in ggplot (default:point_size*0.1).}

\item{stroke_contrast_fac}{Factor by which to darken point outline for contrast (default:0.1).}

\item{grad_palette}{Gradient color palette. one of ("greys", "inferno", "magma", "viridis", "BlGrRd", "RdYlBu", "Spectral") or value to pass as 'colors' argument to 'grDevices::colorRampPalette()'.}

\item{net_attr}{Name of entry in colNets(ace) containing the ACTIONet adjacency matrix to use for value imputation if 'alpha_val>0' (default:'ACTIONet').}

\item{coordinate_attr}{Name of entry in colMaps(ace) containing the 2D plot coordinates (default:'ACTIONet2D').}
}
\value{
'ggplot' object.
}
\description{
Plots gradient of (potentially imputed) values on ACTIONet scatter plot.
}
\examples{
ace = run.ACTIONet(ace)
x = logcounts(ace)['CD14', ]
plot.ACTIONet.gradient(ace, x, trans_attr = ace$node_centrality)
}
