% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.ACTIONet.interactive}
\alias{plot.ACTIONet.interactive}
\title{Main ACTIONet interactive plotting function}
\usage{
\method{plot}{ACTIONet.interactive}(
  ace,
  labels = NULL,
  transparency.attr = NULL,
  trans.z.threshold = -1,
  trans.fact = 1,
  node_size = 1,
  CPal = CPal20,
  enrichment.table = NULL,
  top_features = 7,
  Alt_Text = NULL,
  blacklist_pattern = "\\\\.|^RPL|^RPS|^MRP|^MT-|^MT|MALAT1|B2M|GAPDH",
  threeD = FALSE,
  title = "ACTIONet",
  coordinate_slot = "ACTIONet2D"
)
}
\arguments{
\item{ace}{ACTIONet output object}

\item{labels}{Annotation of interest (clusters, celltypes, etc.) to be projected on the ACTIONet plot}

\item{transparency.attr}{Additional continuous attribute to project onto the transparency of nodes}

\item{trans.z.threshold, }{trans.fact Control the effect of transparency mapping}

\item{node_size}{Size of nodes in the ACTIONet plot}

\item{CPal}{Color palette (named vector or a name for a given known palette)}

\item{enrichment.table}{To project the top-ranked features interactively.}

\item{top_features}{Number of features to show per cell}

\item{blacklist_pattern}{List of genes to filter-out}

\item{threeD}{Whether to show the plot in 3D}

\item{title}{Main title of the plot}
}
\value{
Visualized ACTIONet
}
\description{
Main ACTIONet interactive plotting function
}
\examples{
ace = run.ACTIONet(sce)
plot.ACTIONet.interactive(ace, ace$assigned_archetype)
}
