% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.ACTIONet}
\alias{plot.ACTIONet}
\title{Plot ACTIONet scatter plot}
\usage{
\method{plot}{ACTIONet}(
  data,
  label_attr = NULL,
  color_attr = NULL,
  trans_attr = NULL,
  trans_fac = 1.5,
  trans_th = -0.5,
  point_size = 1,
  stroke_size = point_size * 0.1,
  stroke_contrast_fac = 0.1,
  palette = CPal_default,
  add_text_labels = TRUE,
  text_size = 3,
  nudge_text_labels = FALSE,
  show_legend = FALSE,
  coordinate_attr = "ACTIONet2D",
  color_slot = "denovo_color",
  point_order = NULL,
  use_repel = TRUE
)
}
\arguments{
\item{data}{'ACTIONetExperiment' object or two-column numeric matrix of X-Y coordinates.}

\item{label_attr}{Character vector of length NROW(ace) or colname of 'colData(ace)' containing cell labels of interest (clusters, celltypes, etc.).}

\item{color_attr}{Character vector of length NROW(ace) or colname of 'colData(ace)' containing hex colors to use for each point, or matrix/data.frame containing RGB values to pass to grDevices::rgb().}

\item{trans_attr}{Numeric vector of length NROW(ace) or colname of 'colData(ace)' used to compute point transparency. Smaller values are more transparent.}

\item{trans_fac}{Transparency modifier (default:1.5).}

\item{trans_th}{Minimum Z-score for which points with 'scale(trans_attr) < trans_th' are masked  (default:-0.5).}

\item{point_size}{Size of points in ggplot (default:1).}

\item{stroke_size}{Size of points outline (stroke) in ggplot (default:point_size*0.1).}

\item{stroke_contrast_fac}{Factor by which to darken point outline for contrast (default:0.1).}

\item{palette}{Color palette. character vector of hex colors or a palette name to pass to 'ggpubr::get_palette()').}

\item{add_text_labels}{Whether or not to add floating text labels on the top of annotated clusters. Ignored if 'label_attr=NULL'. (default:'TRUE').}

\item{text_size}{Size of floating text. Passed to 'ggplot2::geom_label()' (default:3).}

\item{nudge_text_labels}{Slightly offset labels proportional to cluster size so as to not cover small clusters (default:'TRUE').}

\item{show_legend}{Show legend based on labels. Ignored if 'label_attr=NULL' or 'color_attr!=NULL' (default:'FALSE').}

\item{coordinate_attr}{Name of entry in colMaps(ace) containing the 2D plot coordinates if 'data' is an 'ACTIONetExperiment' object (default:'ACTIONet2D').}

\item{color_slot}{Name of entry in colMaps(ace) containing RGB values for default point colors generated by 'layout_ACTIONet()'. Used only if no other coloring parameters are given or valid (default:'denovo_color').}

\item{point_order}{Integer vector specifying order in which to plot individual points (default:'NULL').}

\item{use_repel}{Use ggrepel to distribute cluster labels. Known to be buggy if used with RStudio (default:'TRUE').}
}
\value{
'ggplot' object.
}
\description{
Plot ACTIONet scatter plot
}
\examples{
ace = run.ACTIONet(ace)
plot.ACTIONet(ace, ace$assigned_archetype)
}
