% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.individual.gene}
\alias{plot.individual.gene}
\title{Plot gene expression violin plot}
\usage{
\method{plot}{individual.gene}(
  ace,
  labels,
  gene_name,
  features_use = NULL,
  assay_name = "logcounts",
  CPal = CPal20
)
}
\arguments{
\item{ace}{ACTIONet output object}

\item{labels}{Annotation of interest (clusters, celltypes, etc.) to be projected on the ACTIONet plot}

\item{gene_name}{Name of the gene to plot}

\item{CPal}{Color palette (named vector or a name for a given known palette)}
}
\value{
Visualized ACTIONet
}
\description{
Plot gene expression violin plot
}
\examples{
plot.individual.gene(ace, ace$assigned_archetype, 'CD14')
}
