% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.top.k.features}
\alias{plot.top.k.features}
\title{Plots heatmap of the top-ranked features of an enrichment table}
\usage{
\method{plot}{top.k.features}(
  feat_scores,
  top_features = 3,
  normalize = TRUE,
  reorder_columns = TRUE,
  row.title = "Archetypes",
  column.title = "Genes",
  rowPal = "black"
)
}
\arguments{
\item{feat_scores}{An arbitrary enrichment table with columns being archetypes and rows being specificity of features}

\item{top_features}{Number of features to return}

\item{reorder_columns}{Whether to optimally re-order columns of the enrichment table}
}
\value{
Enrichment heatmap
}
\description{
Plots heatmap of the top-ranked features of an enrichment table
}
\examples{
feat_scores = as.matrix(rowMaps(ace)[['unified_feature_specificity']])
plot.top.k.features(feat_scores, 3)
}
