% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{reconstruct.ACTIONet}
\alias{reconstruct.ACTIONet}
\title{Reconstructs the ACTIONet graph with the new parameters (uses prior decomposition)}
\usage{
reconstruct.ACTIONet(
  ace,
  network_density = 1,
  mutual_edges_only = TRUE,
  layout_compactness = 50,
  layout_epochs = 1000,
  thread_no = 0,
  layout_in_parallel = TRUE,
  layout_algorithm = 0,
  reduction_slot = "ACTION",
  output_slot = "ACTIONet",
  seed = 0
)
}
\arguments{
\item{ace}{ACTIONetExperiment object containing the results}

\item{network_density}{Density factor of ACTIONet graph (default=1).}

\item{mutual_edges_only}{Whether to enforce edges to be mutually-nearest-neighbors (default=TRUE).}

\item{thread_no}{Number of parallel threads (default=0)}

\item{reduction_slot}{Slot in the colMaps(ace) that holds reduced kernel (default='S_r')}

\item{compactness_level}{A value between 0-100, indicating the compactness of ACTIONet layout (default=50)}

\item{n_epochs}{Number of epochs for SGD algorithm (default=500).}
}
\value{
ace Updated ace object
}
\description{
Reconstructs the ACTIONet graph with the new parameters (uses prior decomposition)
}
\examples{
plot.ACTIONet(ace)
ace.updated = reconstruct.ACTIONet(ace, network_density = 0.1)
plot.ACTIONet(ace.updated)
}
