% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correct.R
\name{reduce.and.batch.correct.ace.fastMNN}
\alias{reduce.and.batch.correct.ace.fastMNN}
\title{Perform batch correction on `ACTIONetExperiment` and `SingleCellExperiment` objects.}
\usage{
reduce.and.batch.correct.ace.fastMNN(
  ace,
  batch_attr = NULL,
  assay_name = "logcounts",
  reduced_dim = 50,
  MNN_k = 20,
  return_V = FALSE,
  reduction_slot = "MNN",
  V_slot = NULL,
  BPPARAM = SerialParam()
)
}
\description{
Perform batch correction on `ACTIONetExperiment` and `SingleCellExperiment` objects.
}
