% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correct.R
\name{reduce.and.batch.correct.ace.Harmony}
\alias{reduce.and.batch.correct.ace.Harmony}
\title{(It uses Harmony for batch-correction: https://github.com/immunogenomics/harmony)}
\usage{
reduce.and.batch.correct.ace.Harmony(
  ace,
  batch_attr,
  reduced_dim = 50,
  max_iter = 10,
  assay_name = "logcounts",
  reduction_slot = "ACTION",
  seed = 0,
  SVD_algorithm = 0
)
}
\arguments{
\item{ace}{ACTIONetExperiment object}

\item{batch_attr}{Vector of length ncol(ace) or column name of colData(ace) containing batch labels.}

\item{reduced_dim}{Dimension of SVD used for reducing kernel matrix}

\item{max_iter}{Number of SVD iterations}
}
\value{
Reduced ace object with added colMaps(ace)
}
\description{
(It uses Harmony for batch-correction: https://github.com/immunogenomics/harmony)
}
\examples{
ace = import.ace.from.10X(input_path)
batch_attr = ace$Batch # Assumes sample annotations are in the input_path with 'Batch' attribute being provided
ace = reduce.and.batch.correct.ace.Harmony(ace)
}
