% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rerun.layout}
\alias{rerun.layout}
\title{Rerun layout on the ACTIONet graph with new parameters}
\usage{
rerun.layout(
  ace,
  layout_compactness = 50,
  layout_epochs = 1000,
  layout_algorithm = 0,
  thread_no = 0,
  network_density = 1,
  mutual_edges_only = TRUE,
  reduction_slot = "ACTION",
  net_slot = "ACTIONet",
  seed = 0
)
}
\arguments{
\item{ace}{ACTIONetExperiment object containing the results}

\item{layout_compactness}{A value between 0-100, indicating the compactness of ACTIONet layout (default=50)}

\item{layout_epochs}{Number of epochs for SGD algorithm (default=500).}

\item{thread_no}{Number of parallel threads (default=8)}

\item{reduction_slot}{Slot in the colMaps(ace) that holds reduced kernel (default='S_r')}
}
\value{
ace Updated ace object
}
\description{
Rerun layout on the ACTIONet graph with new parameters
}
\examples{
plot.ACTIONet(ace)
ace.updated = rerun.layout(ace, layout_compactness = 20)
plot.ACTIONet(ace.updated)
}
