% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{rowEmbeddings<-}
\alias{rowEmbeddings<-}
\title{Set row embeddings}
\usage{
rowEmbeddings(object, ...) <- value
}
\arguments{
\item{object}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Set row embeddings
}
