% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{rowEmbeddings}
\alias{rowEmbeddings}
\title{Get row embeddings}
\usage{
rowEmbeddings(object, ...)
}
\arguments{
\item{object}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Get row embeddings
}
