% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{rowMapMeta<-}
\alias{rowMapMeta<-}
\title{Set rowMapMeta}
\usage{
rowMapMeta(object, ...) <- value
}
\arguments{
\item{object}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Set rowMapMeta
}
