% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{rowMaps<-}
\alias{rowMaps<-}
\title{Set rowMaps}
\usage{
rowMaps(object, ...) <- value
}
\arguments{
\item{object}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Set rowMaps
}
