% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{rowMapTypes<-}
\alias{rowMapTypes<-}
\title{Set rowMapTypes}
\usage{
rowMapTypes(object, ...) <- value
}
\arguments{
\item{object}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Set rowMapTypes
}
