% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{rowNets<-}
\alias{rowNets<-}
\title{Set rowNets}
\usage{
rowNets(object, ...) <- value
}
\arguments{
\item{object}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Set rowNets
}
