% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run.ACTIONet}
\alias{run.ACTIONet}
\title{A wrapper function to call all main functions of the ACTIONet}
\usage{
run.ACTIONet(
  ace,
  batch = NULL,
  k_min = 2,
  k_max = 40,
  assay_name = "logcounts",
  reduction_slot = "ACTION",
  net_slot_out = "ACTIONet",
  min_cells_per_arch = 2,
  max_iter_ACTION = 100,
  min_specificity_z_thresh = -3,
  network_density = 1,
  mutual_edges_only = TRUE,
  layout_compactness = 50,
  layout_epochs = 1000,
  layout_algorithm = 0,
  layout_in_parallel = TRUE,
  unification_alpha = 0.99,
  unification_sensitivity = 0,
  footprint_alpha = 0.85,
  thread_no = 0,
  full_trace = FALSE,
  seed = 0
)
}
\arguments{
\item{ace}{Reduced `ACTIONetExperiment (ace)` object (output of reduce.ace() function).}

\item{k_max}{Maximum depth of decompositions (default=30).}

\item{assay_name}{Name of assay to be used (default='logcounts').}

\item{reduction_slot}{Slot in the colMaps(ace) that holds reduced kernel (default='S_r').}

\item{min_specificity_z_thresh}{Defines the stringency of pruning nonspecific archetypes.
The larger the value, the more archetypes will be filtered out (default=-1).}

\item{network_density}{Density factor of ACTIONet graph (default=1).}

\item{mutual_edges_only}{Whether to enforce edges to be mutually-nearest-neighbors (default=TRUE).}

\item{layout_compactness}{A value between 0-100, indicating the compactness of ACTIONet layout (default=50).}

\item{layout_epochs}{Number of epochs for SGD algorithm (default=1000).}

\item{thread_no}{Number of parallel threads (default=0).}
}
\value{
A named list: \itemize{
\item ace: ACTIONetExperiment object (derived from SummarizedExperiment)
\item ACTIONet.trace: Log of ACTIONet function calls
}
}
\description{
A wrapper function to call all main functions of the ACTIONet
}
\examples{
ace = reduce(ace)
ACTIONet.out = run.ACTIONet(ace)
ace = ACTIONet.out$ace # main output
trace = ACTIONet.out$trace # for backup
}
