% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCINET_interface.R
\name{run.SCINET.archetype}
\alias{run.SCINET.archetype}
\title{Computes a list of archetype-specific interactomes}
\usage{
run.SCINET.archetype(
  ace,
  G = NULL,
  core = TRUE,
  min.edge.weight = 2,
  spec.sample_no = 1000,
  thread_no = 8,
  compute.topo.specificity = TRUE
)
}
\arguments{
\item{ace}{ACTIONet output object}

\item{G}{Baseline network as an adjacency matrix or igraph object.
If it is NULL, PCNet is loaded as the baseline from SCINET package.}

\item{min.edge.weight}{Used in post-processing to remove insignificant edges}

\item{spec.sample_no}{Number of random samples for computing topological specificity scores}

\item{thread_no}{Number of parallel threads}

\item{compute.topo.specificity}{Whether to compute topological specificity scores (will be stores as V(G)$specificity)}
}
\value{
A list of igraph objects
}
\description{
Computes a list of archetype-specific interactomes
}
\examples{
network.list = run.SCINET.archetype(ace)
G = network.list[[1]]
V(G)$name[order(V(G)$specificity, decreasing = T)[1:10]]
}
