% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCINET_interface.R
\name{run.SCINET.gene.scores}
\alias{run.SCINET.gene.scores}
\title{Computes a list of specific interactomes for an arbitrary gene score matrix}
\usage{
run.SCINET.gene.scores(
  gene.scores,
  G = NULL,
  min.edge.weight = 2,
  spec.sample_no = 1000,
  thread_no = 8,
  compute.topo.specificity = TRUE
)
}
\arguments{
\item{gene.scores}{gene.scores}

\item{G}{Baseline network as an adjacency matrix or igraph object.
If it is NULL, PCNet is loaded as the baseline from SCINET package.}

\item{min.edge.weight}{Used in post-processing to remove insignificant edges}

\item{spec.sample_no}{Number of random samples for computing topological specificity scores}

\item{thread_no}{Number of parallel threads}

\item{compute.topo.specificity}{Whether to compute topological specificity scores (will be stores as V(G)$specificity)}

\item{specificity.slot.name}{An entry in the rowMaps(ace), precomputed using compute.cluster.feature.specificity() function}
}
\value{
A list of igraph objects
}
\description{
Computes a list of specific interactomes for an arbitrary gene score matrix
}
\examples{
network.list = run.SCINET.gene.scores(gene.scores)
G = network.list[[1]]
V(G)$name[order(V(G)$specificity, decreasing = T)[1:10]]
}
