% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_HDBSCAN}
\alias{run_HDBSCAN}
\title{Clusters data points using the hierarchical DBSCAN algorithm.}
\usage{
run_HDBSCAN(X, minPoints = 5L, minClusterSize = 5L)
}
\arguments{
\item{X}{Input data matrix with each row being a data point}
}
\value{
A list with \itemize{
\item labels
\item membershipProbabilities
\item outlierScores
}
}
\description{
Clusters data points using the hierarchical DBSCAN algorithm.
}
\examples{
S_r = t(reducedDims(ace)[["S_r"]])
W_r = S_r \%*\% trace$pruning.out$C_stacked
X = Matrix::t(W_r)
HDBSCAN.out = run_HDBSCAN(X)
clusters = HDBSCAN.out$labels
}
