% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_ace.R
\name{show,ACTIONetExperiment-method}
\alias{show,ACTIONetExperiment-method}
\title{Shows slots of ACTIONetExperiment (ACE) object}
\usage{
\S4method{show}{ACTIONetExperiment}(object)
}
\arguments{
\item{object}{ACTIONetExperiment object}
}
\description{
Shows slots of ACTIONetExperiment (ACE) object
}
