% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ace.R, R/set_ace.R
\name{sizeFactors,ACTIONetExperiment-method}
\alias{sizeFactors,ACTIONetExperiment-method}
\alias{sizeFactors<-,ACTIONetExperiment-method}
\title{Set column-associated size factors}
\usage{
\S4method{sizeFactors}{ACTIONetExperiment}(object)

\S4method{sizeFactors}{ACTIONetExperiment}(object, ...) <- value
}
\description{
Set column-associated size factors
}
