% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_ace.R
\name{[,ACTIONetExperiment,ANY,ANY,ANY-method}
\alias{[,ACTIONetExperiment,ANY,ANY,ANY-method}
\title{Subsets rows/columns of an ACTIONetExperiment (ACE) object}
\usage{
\S4method{[}{ACTIONetExperiment,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{i, j}{rows and columns}
}
\description{
Subsets rows/columns of an ACTIONetExperiment (ACE) object
}
