% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{visualize.markers}
\alias{visualize.markers}
\title{Projects a set of markers on the ACTIONet
It also optionally imputes the markers.}
\usage{
visualize.markers(
  ace,
  markers,
  features_use = NULL,
  assay_name = "logcounts",
  transparency.attr = NULL,
  trans.z.threshold = -0.5,
  trans.fact = 3,
  node_size = 1,
  CPal = "magma",
  alpha_val = 0.9,
  export_path = NA
)
}
\arguments{
\item{ace}{ACTIONet output object}

\item{markers}{Set of row features (e.g. genes) to visualize .}

\item{transparency.attr}{Additional continuous attribute to project onto the transparency of nodes}

\item{trans.z.threshold, }{trans.fact Control the effect of transparency mapping}

\item{node_size}{Size of nodes in the ACTIONet plot}

\item{CPal}{Color palette (named vector or a name for a given known palette)}

\item{alpha_val}{Between [0, 1]. If it is greater than 0, smoothing of scores would be performed}

\item{coordinate_slot}{Entry in colMaps(ace) containing the plot coordinates (default:'ACTIONet2D')}
}
\value{
Visualized ACTIONet with projected scores
}
\description{
Projects a set of markers on the ACTIONet
It also optionally imputes the markers.
}
\examples{
ace = run.ACTIONet(sce)
visualize.markers(ace, markers = c('CD14', 'CD19', 'CD3G'), transparency.attr = ace$node_centrality)
}
